gui/requirements.txt
flask>=3.0
python-frontmatter>=1.1
PyYAML>=6.0
